+++ hutool-core/src/main/java/cn/hutool/core/io/FileTypeUtil.java
@@ -83,6 +83,8 @@ public class FileTypeUtil {
 		FILE_TYPE_MAP.put("AC9EBD8F", "qdf"); // Quicken (qdf)
 		FILE_TYPE_MAP.put("E3828596", "pwl"); // Windows Password (pwl)
 		FILE_TYPE_MAP.put("2E7261FD", "ram"); // Real Audio (ram)
+		// https://stackoverflow.com/questions/45321665/magic-number-for-google-image-format
+		FILE_TYPE_MAP.put("52494646", "webp");
 	}
 
 	/**