+++ hutool-core/src/test/java/cn/hutool/core/io/FileTypeUtilTest.java
@@ -75,4 +75,14 @@ public class FileTypeUtilTest {
 		inputStream.reset();
 	}
 
+	@Test
+	@Ignore
+	public void webpTest(){
+		// https://gitee.com/dromara/hutool/issues/I5BGTF
+		final File file = FileUtil.file("d:/test/a.webp");
+		final BufferedInputStream inputStream = FileUtil.getInputStream(file);
+		final String type = FileTypeUtil.getType(inputStream);
+		Console.log(type);
+	}
+
 }