+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * 【poi    】     PDF相关
 * 【poi    】     HTML、DOCX转换相关
 * 【poi    】     Markdown相关（如HTML转换等），基于commonmark-java
+* 【core   】     针对JDK8+的日期，统一归档到TimeUtil
+* 【core   】     CopyOptions参数重构
+* 【db     】     增加DDL封装
+* 【json   】     删除JSONNull
 
 ### ❌不兼容特性